remove useless cast

+++ proxmox-datacenter-client/src/pve.rs
@@ -705,7 +705,7 @@ async fn remote_migrate_lxc(
         params = params.bwlimit(bwlimit);
     }
     if restart {
-        params = params.restart(true, timeout.map(|secs| Duration::from_secs(secs as u64)));
+        params = params.restart(true, timeout.map(|secs| Duration::from_secs(secs)));
     }
     for mapping in map_storage {
         let pos = mapping
